adding cli flag handling
