+++ config.yml
@@ -4,7 +4,7 @@ AuditEnabled: true
 OverwriteHistoricalAudits: true
 Run: # must be all-lowercase
   - "kubernetes"
-  - "storage"
+  # - "storage"
 ServicePacks:
   Kubernetes:
     KubeConfig: 
@@ -22,7 +22,7 @@ ServicePacks:
       - Name: iam
         Excluded:
       - Name: general
-        Excluded: "out"
+        Excluded:
       - Name: container_registry_access
         Excluded: "out"
       - Name: internet_access
