+++ go.mod
@@ -3,7 +3,7 @@ module github.com/citihub/probr-core
 go 1.14
 
 require (
-	github.com/citihub/probr-sdk v0.0.15
+	github.com/citihub/probr-sdk v0.0.16
 	github.com/hashicorp/go-hclog v0.14.1
 	github.com/hashicorp/go-plugin v1.4.0
 )
